refactor: (#654) 용량이 큰 svg 파일들 webp, png 파일로 변경
