+++ frontend/src/pages/Ranking/PopularPost/index.tsx
@@ -4,9 +4,9 @@ import { usePopularPostRanking } from '@hooks/query/ranking/usePopularPostRankin
 
 import { PATH } from '@constants/path';
 
-import firstRankIcon from '@assets/first-rank.svg';
-import secondRankIcon from '@assets/second-rank.svg';
-import thirdRankIcon from '@assets/third-rank.svg';
+import firstRankIcon from '@assets/first-rank.webp';
+import secondRankIcon from '@assets/second-rank.webp';
+import thirdRankIcon from '@assets/third-rank.webp';
 
 import * as S from './style';
